Added Linux library for compilation
The Linux version of the SFML library was added, the original windows SFML library was renamed with _WIN at the end along with the linux one with _LX at the end instead. It was also added the file serverside-compilation.sh in the server folder to help a linux user compile the file in a more user freinadly way.