Renamed and moved sh script for lx compilation
The sh script for lx compilation was rinamed to make clear its abilty to compile not only the client side binary but also the server side one. Therefore the file was also moved from the client folder to the root folder in order to furtermore spicify its ability to compile also the server code.